pyproject.toml
[tool.poetry]
name = "robo_loader"
version = "0.1.0"
description = ""
authors = ["Semih Koyuncu"]


[tool.poetry.dependencies]
python = ">=3.12,<3.14"
loguru = "^0.7.2"
virtualenv = "^20.27.1"
pydrive = "^1.3.1"
rich = "^13.9.4"
pyserial = "^3.5"
fastapi = {extras = ["standard"], version = "^0.115.4"}
numpy = "^2.1.3"
wxwidgets = "^1.0.5"
pygame = "^2.6.1"
sounddevice = "^0.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
